second session assigment

+++ src/package1/Sum.java
@@ -3,8 +3,8 @@ package package1;
 public class Sum {
 	
 	
-	public static int sumResult(int a,int b) {
-		int result;
+	public static double sumResult(double a,double b) {
+		double result;
 		
 		result=a+b;
 		return result;
